Avoid setting a non SELECT sql to be a SELECT
If the sql query contains SELECT word any where inside it then we are mistaking it to be a SELECT query. Thus we mistake something like this
Insert into table1 (`student_name`) values ('select') to be a SELECT query. 
Thence we don't try to get back the generated_key from the db.